+++ connection/connection.js
@@ -5,7 +5,7 @@ import mongoose from 'mongoose';
 
 const databaseConnection = async () =>{
     try {
-        await mongoose.connect(process.env.CONNECTIONSTRING , {useNewUrlParser: true , useUnifiedTopology: true });
+        // await mongoose.connect(process.env.CONNECTIONSTRING , {useNewUrlParser: true , useUnifiedTopology: true });
         // console.log('Connected');
     } catch (error) {
         console.log('Database can be connected' , error.message);
