+++ functions/sendEmail.js
@@ -3,33 +3,25 @@ import nodemailer  from 'nodemailer';
 import dotEnv from 'dotenv';
 dotEnv.config();
 const sendEmail = async (toEmail) => {
-    // console.log(process.env.EMAIL_PASSWORD);
-    // create reusable transporter object using the default SMTP transport
     let transporter = nodemailer.createTransport({
       host: "smtp.gmail.com",
       port: 587,
-      secure: false, // true for 465, false for other ports
+      secure: false,
       auth: {
-        user: "codeinblog123@gmail.com", // generated ethereal user
-        pass: process.env.EMAIL_PASSWORD, // generated ethereal password
+        user: "codeinblog123@gmail.com", 
+        pass: process.env.EMAIL_PASSWORD, 
       },
     });
-  
-    // send mail with defined transport object
     let info = await transporter.sendMail({
-      from: '"Fred Foo 👻" <foo@example.com>', // sender address
-      to: toEmail, // list of receivers
-      subject: "Hello ✔", // Subject line
-      text: "Hello world?", // plain text body
-      html: "<b>Hello world?</b>", // html body
+      from: 'Sezerano J Chrysostome 👻" <codeinblog@gmail.com>', 
+      to: toEmail, 
+      subject: "Hello ✔ please checkout this blog", 
+      text: "Hello world?", 
+      html: "<b>Hello world?</b>", 
     });
   
-  //   console.log("Message sent: %s", info.messageId);
-    // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
+    // console.log("Message sent: %s", info.messageId);
   
-      // Preview only available when sending through an Ethereal account
-     //  console.log("Preview URL: %s", nodemailer.getTestMessageUrl(info));
-    // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
   }
   
   export default sendEmail ;
